requirements.txt
Flask>=2.2
python-dotenv>=1.0
flake8>=6.0
black>=23.0